Avoid casting issue on Linux/GCC

+++ source/blender/blenlib/intern/math_bits_inline.cc
@@ -52,7 +52,7 @@ MINLINE unsigned int bitscan_forward_clear_uint(unsigned int *a)
   return i;
 }
 
-MINLINE unsigned int bitscan_forward_clear_uint64(unsigned long long *a)
+MINLINE unsigned int bitscan_forward_clear_uint64(uint64_t *a)
 {
   unsigned int i = bitscan_forward_uint64(*a);
   *a &= (*a) - 1;
